refactor: 拆分 Core storage session/memory 长文件 helper (#115)
