+++ qq-maid-core/src/storage/memory/mod.rs
@@ -5,9 +5,7 @@
 //! 也不保留 JSONL 回退，避免写入失败时出现“表面成功、实际未保存”的状态。
 
 use qq_maid_common::redaction::redact_sensitive_text;
-use rusqlite::{
-    Connection, OptionalExtension, Row, params, params_from_iter, types::Value as SqlValue,
-};
+use rusqlite::{Connection, params};
 use serde::{Deserialize, Serialize};
 use uuid::Uuid;
 
@@ -15,6 +13,24 @@ use crate::{
     storage::database::{DatabaseError, SqliteDatabase, SqliteMigration},
     util::time_context::now_iso_cn,
 };
+
+// 拆分出的纯 helper 子模块：均不改变 schema、权限语义与已确认持久化格式。
+mod clean;
+mod permission;
+mod query;
+mod row;
+
+use clean::{
+    clean_optional, clean_optional_option, clean_required, clean_scope_id, default_memory_type,
+    default_scope, infer_legacy_scope_identity, legacy_unassigned_scope_type,
+};
+use permission::ensure_can_modify;
+use query::{
+    apply_update_to_record, list_accessible_for_context_unlocked, list_scoped_unlocked,
+    list_unlocked, resolve_memory_id_scoped_unlocked, resolve_memory_id_unlocked,
+    update_record_unlocked,
+};
+use row::{get_by_id_scoped_unlocked, get_by_id_unlocked};
 
 /// Memory schema migration，由应用启动时的通用数据库初始化流程统一执行。
 ///
@@ -640,445 +656,6 @@ impl UpdateMemoryRequest {
             || self.memory_type.is_some()
             || self.scope.is_some()
     }
-}
-
-fn list_unlocked(
-    conn: &Connection,
-    query: &ListMemoryQuery,
-) -> Result<Vec<MemoryRecord>, MemoryError> {
-    let mut sql = String::from(
-        "SELECT id, created_at, updated_at, memory_type, scope,
-                scope_type, scope_id, created_by_user_id,
-                user_id, group_id, content, source_text
-         FROM memories
-         WHERE 1 = 1",
-    );
-    let mut values = Vec::<SqlValue>::new();
-
-    push_optional_filter(
-        &mut sql,
-        &mut values,
-        "scope",
-        clean_optional_option(query.scope.clone()),
-    );
-    push_optional_filter(
-        &mut sql,
-        &mut values,
-        "memory_type",
-        clean_optional_option(query.memory_type.clone()),
-    );
-    push_optional_filter(
-        &mut sql,
-        &mut values,
-        "user_id",
-        clean_optional_option(query.user_id.clone()),
-    );
-    push_optional_filter(
-        &mut sql,
-        &mut values,
-        "group_id",
-        clean_optional_option(query.group_id.clone()),
-    );
-    if let Some(q) = clean_optional_option(query.q.clone()) {
-        sql.push_str(
-            " AND (instr(lower(content), lower(?)) > 0 OR instr(lower(source_text), lower(?)) > 0)",
-        );
-        values.push(SqlValue::Text(q.clone()));
-        values.push(SqlValue::Text(q));
-    }
-    sql.push_str(" ORDER BY row_id DESC LIMIT ?");
-    values.push(SqlValue::Integer(query.limit() as i64));
-
-    let mut stmt = conn.prepare(&sql).map_err(MemoryError::from_sql)?;
-    let rows = stmt
-        .query_map(params_from_iter(values.iter()), memory_from_row)
-        .map_err(MemoryError::from_sql)?;
-    collect_rows(rows)
-}
-
-fn list_scoped_unlocked(
-    conn: &Connection,
-    query: &ScopedMemoryQuery,
-) -> Result<Vec<MemoryRecord>, MemoryError> {
-    let scope_id = clean_scope_id(&query.scope_id)?;
-    let mut sql = String::from(
-        "SELECT id, created_at, updated_at, memory_type, scope,
-                scope_type, scope_id, created_by_user_id,
-                user_id, group_id, content, source_text
-         FROM memories
-         WHERE scope_type = ? AND scope_id = ?",
-    );
-    let mut values = vec![
-        SqlValue::Text(query.scope_type.as_str().to_owned()),
-        SqlValue::Text(scope_id),
-    ];
-
-    push_optional_filter(
-        &mut sql,
-        &mut values,
-        "scope",
-        clean_optional_option(query.scope.clone()),
-    );
-    push_optional_filter(
-        &mut sql,
-        &mut values,
-        "memory_type",
-        clean_optional_option(query.memory_type.clone()),
-    );
-    if let Some(q) = clean_optional_option(query.q.clone()) {
-        sql.push_str(
-            " AND (instr(lower(content), lower(?)) > 0 OR instr(lower(source_text), lower(?)) > 0)",
-        );
-        values.push(SqlValue::Text(q.clone()));
-        values.push(SqlValue::Text(q));
-    }
-    sql.push_str(" ORDER BY row_id DESC LIMIT ?");
-    values.push(SqlValue::Integer(query.limit() as i64));
-
-    let mut stmt = conn.prepare(&sql).map_err(MemoryError::from_sql)?;
-    let rows = stmt
-        .query_map(params_from_iter(values.iter()), memory_from_row)
-        .map_err(MemoryError::from_sql)?;
-    collect_rows(rows)
-}
-
-fn list_accessible_for_context_unlocked(
-    conn: &Connection,
-    personal_scope_id: Option<&str>,
-    group_scope_id: Option<&str>,
-    limit: usize,
-) -> Result<Vec<MemoryRecord>, MemoryError> {
-    let mut clauses = Vec::new();
-    let mut values = Vec::<SqlValue>::new();
-    if let Some(scope_id) = personal_scope_id.and_then(clean_optional_str) {
-        clauses.push("(scope_type = ? AND scope_id = ?)");
-        values.push(SqlValue::Text(
-            MemoryScopeType::Personal.as_str().to_owned(),
-        ));
-        values.push(SqlValue::Text(scope_id));
-    }
-    if let Some(scope_id) = group_scope_id.and_then(clean_optional_str) {
-        clauses.push("(scope_type = ? AND scope_id = ?)");
-        values.push(SqlValue::Text(MemoryScopeType::Group.as_str().to_owned()));
-        values.push(SqlValue::Text(scope_id));
-    }
-    if clauses.is_empty() {
-        return Ok(Vec::new());
-    }
-
-    let sql = format!(
-        "SELECT id, created_at, updated_at, memory_type, scope,
-                scope_type, scope_id, created_by_user_id,
-                user_id, group_id, content, source_text
-         FROM memories
-         WHERE {}
-         ORDER BY row_id DESC
-         LIMIT ?",
-        clauses.join(" OR ")
-    );
-    values.push(SqlValue::Integer(limit.clamp(1, 100) as i64));
-
-    let mut stmt = conn.prepare(&sql).map_err(MemoryError::from_sql)?;
-    let rows = stmt
-        .query_map(params_from_iter(values.iter()), memory_from_row)
-        .map_err(MemoryError::from_sql)?;
-    collect_rows(rows)
-}
-
-fn push_optional_filter(
-    sql: &mut String,
-    values: &mut Vec<SqlValue>,
-    column: &str,
-    value: Option<String>,
-) {
-    if let Some(value) = value {
-        sql.push_str(" AND ");
-        sql.push_str(column);
-        sql.push_str(" = ?");
-        values.push(SqlValue::Text(value));
-    }
-}
-
-/// 根据完整 ID 或前缀解析真实 ID。
-/// 前缀至少需要 4 个字符，且不能有多条匹配。
-fn resolve_memory_id_unlocked(
-    conn: &Connection,
-    id_or_prefix: &str,
-) -> Result<String, MemoryError> {
-    let target = id_or_prefix.trim();
-    if target.is_empty() {
-        return Err(MemoryError::bad_request("memory id is required"));
-    }
-
-    if let Some(id) = conn
-        .query_row(
-            "SELECT id FROM memories WHERE id = ?1",
-            params![target],
-            |row| row.get::<_, String>(0),
-        )
-        .optional()
-        .map_err(MemoryError::from_sql)?
-    {
-        return Ok(id);
-    }
-
-    if target.chars().count() < 4 {
-        return Err(MemoryError::bad_request(
-            "memory id prefix must contain at least 4 characters",
-        ));
-    }
-
-    let mut stmt = conn
-        .prepare(
-            "SELECT id
-             FROM memories
-             WHERE substr(id, 1, length(?1)) = ?1
-             ORDER BY row_id DESC
-             LIMIT 2",
-        )
-        .map_err(MemoryError::from_sql)?;
-    let rows = stmt
-        .query_map(params![target], |row| row.get::<_, String>(0))
-        .map_err(MemoryError::from_sql)?;
-    let matches = rows
-        .collect::<Result<Vec<_>, _>>()
-        .map_err(MemoryError::from_sql)?;
-
-    match matches.as_slice() {
-        [id] => Ok(id.clone()),
-        [] => Err(MemoryError::not_found("memory not found")),
-        _ => Err(MemoryError::bad_request("memory id prefix is ambiguous")),
-    }
-}
-
-fn resolve_memory_id_scoped_unlocked(
-    conn: &Connection,
-    scope_type: MemoryScopeType,
-    scope_id: &str,
-    id_or_prefix: &str,
-) -> Result<String, MemoryError> {
-    let scope_id = clean_scope_id(scope_id)?;
-    let target = id_or_prefix.trim();
-    if target.is_empty() {
-        return Err(MemoryError::bad_request("memory id is required"));
-    }
-
-    if let Some(id) = conn
-        .query_row(
-            "SELECT id FROM memories
-             WHERE id = ?1 AND scope_type = ?2 AND scope_id = ?3",
-            params![target, scope_type.as_str(), scope_id.as_str()],
-            |row| row.get::<_, String>(0),
-        )
-        .optional()
-        .map_err(MemoryError::from_sql)?
-    {
-        return Ok(id);
-    }
-
-    if target.chars().count() < 4 {
-        return Err(MemoryError::bad_request(
-            "memory id prefix must contain at least 4 characters",
-        ));
-    }
-
-    let mut stmt = conn
-        .prepare(
-            "SELECT id
-             FROM memories
-             WHERE scope_type = ?1
-               AND scope_id = ?2
-               AND substr(id, 1, length(?3)) = ?3
-             ORDER BY row_id DESC
-             LIMIT 2",
-        )
-        .map_err(MemoryError::from_sql)?;
-    let rows = stmt
-        .query_map(
-            params![scope_type.as_str(), scope_id.as_str(), target],
-            |row| row.get::<_, String>(0),
-        )
-        .map_err(MemoryError::from_sql)?;
-    let matches = rows
-        .collect::<Result<Vec<_>, _>>()
-        .map_err(MemoryError::from_sql)?;
-
-    match matches.as_slice() {
-        [id] => Ok(id.clone()),
-        [] => Err(MemoryError::not_found("memory not found")),
-        _ => Err(MemoryError::bad_request("memory id prefix is ambiguous")),
-    }
-}
-
-fn get_by_id_unlocked(conn: &Connection, id: &str) -> Result<Option<MemoryRecord>, MemoryError> {
-    conn.query_row(
-        "SELECT id, created_at, updated_at, memory_type, scope,
-                scope_type, scope_id, created_by_user_id,
-                user_id, group_id, content, source_text
-         FROM memories
-         WHERE id = ?1",
-        params![id],
-        memory_from_row,
-    )
-    .optional()
-    .map_err(MemoryError::from_sql)
-}
-
-fn get_by_id_scoped_unlocked(
-    conn: &Connection,
-    scope_type: MemoryScopeType,
-    scope_id: &str,
-    id: &str,
-) -> Result<Option<MemoryRecord>, MemoryError> {
-    let scope_id = clean_scope_id(scope_id)?;
-    conn.query_row(
-        "SELECT id, created_at, updated_at, memory_type, scope,
-                scope_type, scope_id, created_by_user_id,
-                user_id, group_id, content, source_text
-         FROM memories
-         WHERE id = ?1 AND scope_type = ?2 AND scope_id = ?3",
-        params![id, scope_type.as_str(), scope_id],
-        memory_from_row,
-    )
-    .optional()
-    .map_err(MemoryError::from_sql)
-}
-
-fn memory_from_row(row: &Row<'_>) -> rusqlite::Result<MemoryRecord> {
-    let created_at: String = row.get(1)?;
-    Ok(MemoryRecord {
-        id: row.get(0)?,
-        ts: created_at.clone(),
-        created_at,
-        updated_at: row.get(2)?,
-        memory_type: row.get(3)?,
-        scope: row.get(4)?,
-        scope_type: row.get(5)?,
-        scope_id: row.get(6)?,
-        created_by_user_id: row.get(7)?,
-        user_id: row.get(8)?,
-        group_id: row.get(9)?,
-        content: row.get(10)?,
-        source_text: row.get(11)?,
-    })
-}
-
-fn apply_update_to_record(
-    record: &mut MemoryRecord,
-    req: UpdateMemoryRequest,
-) -> Result<(), MemoryError> {
-    if let Some(content) = req.content {
-        record.content = redact_sensitive_text(&clean_required(content, "content")?);
-    }
-    if let Some(source_text) = req.source_text {
-        record.source_text = redact_sensitive_text(&source_text);
-    }
-    if let Some(memory_type) = req.memory_type.and_then(clean_optional) {
-        record.memory_type = memory_type;
-    }
-    if let Some(scope) = req.scope.and_then(clean_optional) {
-        record.scope = scope;
-    }
-    record.updated_at = Some(now_iso_cn());
-    Ok(())
-}
-
-fn update_record_unlocked(conn: &Connection, record: &MemoryRecord) -> Result<(), MemoryError> {
-    conn.execute(
-        "UPDATE memories
-         SET content = ?1, source_text = ?2, memory_type = ?3, scope = ?4, updated_at = ?5
-         WHERE id = ?6",
-        params![
-            record.content.as_str(),
-            record.source_text.as_str(),
-            record.memory_type.as_str(),
-            record.scope.as_str(),
-            record.updated_at.as_deref(),
-            record.id.as_str(),
-        ],
-    )
-    .map_err(MemoryError::from_sql)?;
-    Ok(())
-}
-
-fn ensure_can_modify(record: &MemoryRecord, actor: &MemoryActor) -> Result<(), MemoryError> {
-    // 群记忆第一版没有管理员识别能力，只允许创建者修改/删除；
-    // created_by_user_id 为空的历史群记忆可读可注入，但不可被普通成员管理。
-    if record.scope_type == MemoryScopeType::Group.as_str()
-        && record.created_by_user_id.as_deref() != Some(actor.user_id.as_str())
-    {
-        return Err(MemoryError::forbidden(
-            "memory is not editable in this scope",
-        ));
-    }
-    Ok(())
-}
-
-fn infer_legacy_scope_identity(
-    user_id: Option<&str>,
-    group_id: Option<&str>,
-) -> (MemoryScopeType, String, String) {
-    if let Some(user_id) = user_id.and_then(clean_optional_str) {
-        return (MemoryScopeType::Personal, user_id.clone(), user_id);
-    }
-    if let Some(group_id) = group_id.and_then(clean_optional_str) {
-        return (
-            MemoryScopeType::Group,
-            group_id,
-            "legacy_unknown_user".to_owned(),
-        );
-    }
-    (
-        MemoryScopeType::LegacyUnassigned,
-        "legacy_unassigned".to_owned(),
-        "legacy_unknown_user".to_owned(),
-    )
-}
-
-fn clean_scope_id(value: &str) -> Result<String, MemoryError> {
-    clean_optional(value.to_owned()).ok_or_else(|| MemoryError::bad_request("scope_id is required"))
-}
-
-fn collect_rows<T, F>(rows: rusqlite::MappedRows<'_, F>) -> Result<Vec<T>, MemoryError>
-where
-    F: FnMut(&Row<'_>) -> rusqlite::Result<T>,
-{
-    rows.collect::<Result<Vec<_>, _>>()
-        .map_err(MemoryError::from_sql)
-}
-
-/// 清理并验证必填字段：去除首尾空格，空值则返回错误。
-fn clean_required(value: String, field: &str) -> Result<String, MemoryError> {
-    clean_optional(value).ok_or_else(|| MemoryError::bad_request(format!("{field} is required")))
-}
-
-/// 清理可选字段：去除首尾空格，空值返回 None。
-fn clean_optional(value: String) -> Option<String> {
-    let value = value.trim().to_owned();
-    if value.is_empty() { None } else { Some(value) }
-}
-
-fn clean_optional_str(value: &str) -> Option<String> {
-    clean_optional(value.to_owned())
-}
-
-/// 清理可选 Option 字段：内层值空则返回 None。
-fn clean_optional_option(value: Option<String>) -> Option<String> {
-    value.and_then(clean_optional)
-}
-
-/// 默认记忆类型。
-fn default_memory_type() -> String {
-    "note".to_owned()
-}
-
-/// 默认记忆作用域。
-fn default_scope() -> String {
-    "general".to_owned()
-}
-
-fn legacy_unassigned_scope_type() -> String {
-    MemoryScopeType::LegacyUnassigned.as_str().to_owned()
 }
 
 #[cfg(test)]
