+++ qq-maid-core/src/storage/session/tests.rs
@@ -1,5 +1,6 @@
 use super::*;
 use crate::runtime::pending::PendingMemory;
+use uuid::Uuid;
 
 fn test_store() -> SessionStore {
     SessionStore::new(
